mudança de js para react
cadastro de usuário mudado para react, que antes estava como js
e alterações menores para funcionamento.
@DouglasBezerra01
