+++ AlfaEduca/package.json
@@ -14,9 +14,9 @@
     "@expo/metro-runtime": "~4.0.0",
     "AlfaEduca": "file:",
     "axios": "^1.7.7",
-    "depcheck": "^0.9.0",
+    "depcheck": "^1.4.7",
     "expo": "~52.0.8",
-    "expo-app-loading": "^2.1.1",
+    "expo-app-loading": "^1.0.3",
     "expo-camera": "~16.0.7",
     "expo-font": "~13.0.1",
     "expo-image-picker": "~16.0.3",
